suppressed binary_fill_holes -> segment
